requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
markdown==3.5.1
